requirements.txt
gradio
ollama
mcp[cli]
langchain
langchain-mcp-adapters
langchain-ollama
langgraph
langchain-community
langgraph-swarm
python-dotenv
PyPDF2
python-docx
python-pptx
fastapi
uvicorn
fastapi-mcp
SQLAlchemy
openpyxl
a2a-sdk
streamlit
langgraph-supervisor
